Move ber archetype creation to estimate dir
So keep cleaning of ber and arch gen separate
