+++ src/drem/transform/ber.py
@@ -1,7 +1,6 @@
 import numpy as np
 import pandas as pd
 
-from icontract import require
 from prefect import task
 
 
@@ -46,20 +45,6 @@ def _bin_year_of_construction_as_in_census(ber: pd.DataFrame) -> pd.DataFrame:
     return ber
 
 
-@require(
-    lambda ber: set(ber.columns)
-    == {"postcodes", "period_built", "total_heat_demand_per_hh"},
-)
-def _get_mean_total_heat_demand_per_hh(ber: pd.DataFrame) -> pd.DataFrame:
-
-    return (
-        ber.groupby(["postcodes", "period_built"])["total_heat_demand_per_hh"]
-        .mean()
-        .rename("mean_heat_demand_per_hh")
-        .reset_index()
-    )
-
-
 @task
 def transform_ber(ber_raw: pd.DataFrame) -> pd.DataFrame:
     """Tidy BER data set.
@@ -73,30 +58,7 @@ def transform_ber(ber_raw: pd.DataFrame) -> pd.DataFrame:
         pd.DataFrame: Clean Dublin BER Residential Housing Data set
     """
     return (
-        ber_raw.loc[
-            :,
-            [
-                "CountyName",
-                "Year_of_Construction",
-                "DeliveredEnergyMainSpace",
-                "DeliveredEnergyMainWater",
-                "DeliveredEnergySecondarySpace",
-                "DeliveredEnergySupplementaryWater",
-            ],
-        ]
-        .pipe(_extract_dublin_rows)
+        ber_raw.pipe(_extract_dublin_rows)
         .pipe(_bin_year_of_construction_as_in_census)
-        .assign(
-            total_heat_demand_per_hh=lambda x: x[
-                [
-                    "DeliveredEnergyMainSpace",
-                    "DeliveredEnergyMainWater",
-                    "DeliveredEnergySecondarySpace",
-                    "DeliveredEnergySupplementaryWater",
-                ]
-            ].sum(axis=1),
-        )
         .rename(columns={"CountyName": "postcodes"})
-        .loc[:, ["postcodes", "period_built", "total_heat_demand_per_hh"]]
-        .pipe(_get_mean_total_heat_demand_per_hh)
     )
